move movie components to new movie folder and update imports
