+++ client/src/components/movies/AllMovies.tsx
