+++ client/src/pages/IndexPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { MovieCards } from "../components/MovieCards";
+import { MovieCards } from "../components/movies/MovieCards";
 import { Login } from "../components/registration/Login";
 
 export const IndexPage: React.FC = () => {
